Had create-tests.js refactored.  Plenty more potential especially between create-tests and run-tests but pausing such efforts for now.
